move pat::XGBooster out of PhysicsTools/PatAlgos into its own package
