+++ PhysicsTools/PatAlgos/BuildFile.xml
@@ -16,6 +16,7 @@
 <use name="PhysicsTools/PatUtils"/>
 <use name="PhysicsTools/TensorFlow"/>
 <use name="PhysicsTools/ONNXRuntime"/>
+<use name="PhysicsTools/XGBoost"/>
 <use name="TrackingTools/Records"/>
 <use name="clhep"/>
 <use name="xgboost"/>
