+++ PhysicsTools/PatAlgos/plugins/PATMuonProducer.cc
@@ -47,8 +47,8 @@
 #include "PhysicsTools/PatAlgos/interface/PATUserDataHelper.h"
 #include "PhysicsTools/PatAlgos/interface/SoftMuonMvaEstimator.h"
 #include "PhysicsTools/PatAlgos/interface/SoftMuonMvaRun3Estimator.h"
-#include "PhysicsTools/PatAlgos/interface/XGBooster.h"
 #include "PhysicsTools/PatUtils/interface/MiniIsolation.h"
+#include "PhysicsTools/XGBoost/interface/XGBooster.h"
 #include "TrackingTools/IPTools/interface/IPTools.h"
 #include "TrackingTools/Records/interface/TransientTrackRecord.h"
 #include "TrackingTools/TransientTrack/interface/TransientTrack.h"
