+++ PhysicsTools/XGBoost/BuildFile.xml
@@ -1,0 +1,4 @@
+<use name="xgboost"/>
+<export>
+  <lib name="1"/>
+</export>
