+++ PhysicsTools/XGBoost/interface/XGBooster.h
@@ -1,5 +1,5 @@
-#ifndef PhysicsTools_PatAlgos_XGBooster_h
-#define PhysicsTools_PatAlgos_XGBooster_h
+#ifndef PhysicsTools_XGBoost_XGBooster_h
+#define PhysicsTools_XGBoost_XGBooster_h
 
 #include <memory>
 #include <string>
